Mapgen variants, guns and ammo expansion
* Added a variant of the M60 Patton that's partially disarmed for
display or storage. Spawns without extra items, secondary armament, or
fuel. A few locations use them.
* Added mapgen variants for all standard garages and the gas station
garage location. In these variants you can find signs of military
presence and a military vehicle, usually an M60 Patton. This Patton will
sometimes be fully kitted out, sometimes the mothballed variant,
implication is that some official had sent an old tank out of storage in
an attempt to rearm them for emergency armor support.
* Added a mapgen variant to the military outpost, effectively a small
armor depot. Bulk of the space is used for a garage that can spawn a
military vehicle.
* Added a cut-down 25mm cannon, in the same vein as the .50 caliber
rifle made from an M2 Browning. In this case however, the end result is
still so heavy that the benefit gained is marginal, mostly just the
ability to use it as a mounted gun without installing it on a vehicle.
You can still choose to slap it onto a vehicle if you want down, as it
has a vehiclepart variant too.
* Added blackpowder variants for the various ammunition handloads. Uses
the standard stat variant blackpowder ammo tends to use.
* The various single-shot main guns now gain some level of blackpowder
tolerance.
* Minor stat tweaks to the standard 25mm autocannon.
* Massively beefed up recoil. Recoil numbers are always quite arbitrary,
but went with the weapon's implied muzzle energy in foot-pounds.
* Added IRREPLACEABLE_CONSUMABLE to the standard ammunition.
* String names for items.
* Belatedly fixed 105mm rounds not having uncrafts.
* Obsoleted the 105mm shotshells because the 105mm gun is rifled.
* Removed loudness values that were assigned to the guns themselves for
some wacky reason.