Some updates
* Added a basic scenario mission for Armored Apocalypse to find a source of fuel for the tank. Fixes https://github.com/chaosvolt/cdda-tankmod-revived-mod/issues/12
* Per feedback, shaved the fake barrel length of the M1 Abrams by a tile, added the hatch linking the driver's postion to the turret (IRL you wouldn't be using this without the gun facing backwards, but oh well), and moved the ammo haphazardly stowed on a rack outside into additional storage the gunner can now reach.
* Also added some added spawns of personal items in the now-empty external rack.

Two problems I was unable to fix. One, the M60 lacks any real space to axe external ammo storage without some fiddling with lower-tier ammo storage items. Two, there's no way to implement a control setup to let you fire the main gun from the commander's hatch without also being able to drive the tank from there, something that is not in fact a thing IRL on the Abrams (but aiming the turret from that position is).

I'd also like to model the bottom hatch on the M60 Patton but good fucking luck figuring out how to do so on a 2D-only vehicle floorplan.