Allow using 25mm dismounted
Per feedback this permits treating the 25mm Bushmaster and makeshift version like heavy weapon such as the M2 or Mk. 19, allowing you to reload it and just use it on foot without exploiting code jank.

Will also pair well with https://github.com/cataclysmbnteam/Cataclysm-BN/pull/3691 for power armor lovers in BN version but not dependant on it.